@@ -1,6 +1,6 @@
 module github.com/reeflective/readline
 
-go 1.20
+go 1.21
 
 require (
 	github.com/google/go-cmp v0.5.8
