@@ -3,6 +3,7 @@ package completion
 import (
 	"bufio"
 	"fmt"
+	"regexp"
 	"strings"
 
 	"github.com/reeflective/readline/internal/color"
@@ -30,7 +31,7 @@ func Display(eng *Engine, maxRows int) {
 	completions := term.ClearLineAfter
 
 	for _, group := range eng.groups {
-		completions += group.writeComps(eng)
+		completions += eng.renderCompletions(group)
 	}
 
 	// Crop the completions so that it fits within our terminal
@@ -45,6 +46,129 @@ func Display(eng *Engine, maxRows int) {
 // when displaying the completions with Display().
 func Coordinates(e *Engine) int {
 	return e.usedY
+}
+
+// renderCompletions renders all completions in a given list (with aliases or not).
+// The descriptions list argument is optional.
+func (e *Engine) renderCompletions(grp *group) string {
+	var builder strings.Builder
+
+	if len(grp.rows) == 0 {
+		return ""
+	}
+
+	if grp.tag != "" {
+		tag := fmt.Sprintf("%s%s%s %s", color.Bold, color.FgYellow, grp.tag, color.Reset)
+		builder.WriteString(tag + term.ClearLineAfter + term.NewlineReturn)
+	}
+
+	for rowIndex, row := range grp.rows {
+		for columnIndex := range grp.columnsWidth {
+			var value Candidate
+
+			// If there are aliases, we might have no completions at the current
+			// coordinates, so just print the corresponding padding and return.
+			if len(row) > columnIndex {
+				value = row[columnIndex]
+			}
+
+			// Apply all highlightings to the displayed value:
+			// selection, prefixes, styles and other things,
+			padding := grp.getPad(value, columnIndex, false)
+			isSelected := rowIndex == grp.posY && columnIndex == grp.posX && grp.isCurrent
+			display := e.highlightDisplay(grp, value, padding, columnIndex, isSelected)
+
+			builder.WriteString(display)
+
+			// Add description if no aliases, or if done with them.
+			onLast := columnIndex == len(grp.columnsWidth)-1
+			if grp.aliased && onLast && value.Description == "" {
+				value = row[0]
+			}
+
+			if !grp.aliased || onLast {
+				grp.maxDescAllowed = grp.setMaximumSizes(columnIndex)
+
+				descPad := grp.getPad(value, columnIndex, true)
+				desc := e.highlightDesc(grp, value, descPad, rowIndex, columnIndex, isSelected)
+				builder.WriteString(desc)
+			}
+		}
+
+		// We're done for this line.
+		builder.WriteString(term.ClearLineAfter + term.NewlineReturn)
+	}
+
+	return builder.String()
+}
+
+func (e *Engine) highlightDisplay(grp *group, val Candidate, pad, col int, selected bool) (candidate string) {
+	// An empty display value means padding.
+	if val.Display == "" {
+		return padSpace(pad)
+	}
+
+	reset := color.Fmt(val.Style)
+	candidate, padded := grp.trimDisplay(val, pad, col)
+
+	if e.IsearchRegex != nil && e.isearchBuf.Len() > 0 && !selected {
+		match := e.IsearchRegex.FindString(candidate)
+		match = color.Fmt(color.Bg+"244") + match + color.Reset + reset
+		candidate = e.IsearchRegex.ReplaceAllLiteralString(candidate, match)
+	}
+
+	if selected {
+		// If the comp is currently selected, overwrite any highlighting already applied.
+		userStyle := color.UnquoteRC(e.config.GetString("completion-selection-style"))
+		selectionHighlightStyle := color.Fmt(color.Bg+"255") + userStyle
+		candidate = selectionHighlightStyle + candidate
+
+		if grp.aliased {
+			candidate += color.Reset
+		}
+	} else {
+		// Highlight the prefix if any and configured for it.
+		if e.config.GetBool("colored-completion-prefix") && e.prefix != "" {
+			if prefixMatch, err := regexp.Compile(fmt.Sprintf("^%s", e.prefix)); err == nil {
+				prefixColored := color.Bold + color.FgBlue + e.prefix + color.BoldReset + color.FgDefault + reset
+				candidate = prefixMatch.ReplaceAllString(candidate, prefixColored)
+			}
+		}
+
+		candidate = reset + candidate + color.Reset
+	}
+
+	return candidate + padded
+}
+
+func (e *Engine) highlightDesc(grp *group, val Candidate, pad, row, col int, selected bool) (desc string) {
+	if val.Description == "" {
+		return color.Reset
+	}
+
+	desc, padded := grp.trimDesc(val, pad)
+
+	// If the next row has the same completions, replace the description with our hint.
+	if len(grp.rows) > row+1 && grp.rows[row+1][0].Description == val.Description {
+		desc = "|"
+	} else if e.IsearchRegex != nil && e.isearchBuf.Len() > 0 && !selected {
+		match := e.IsearchRegex.FindString(desc)
+		match = color.Fmt(color.Bg+"244") + match + color.Reset + color.Dim
+		desc = e.IsearchRegex.ReplaceAllLiteralString(desc, match)
+	}
+
+	// If the comp is currently selected, overwrite any highlighting already applied.
+	// Replace all background reset escape sequences in it, to ensure correct display.
+	if row == grp.posY && col == grp.posX && grp.isCurrent && !grp.aliased {
+		userDescStyle := color.UnquoteRC(e.config.GetString("completion-selection-style"))
+		selectionHighlightStyle := color.Fmt(color.Bg+"255") + userDescStyle
+		desc = strings.ReplaceAll(desc, color.BgDefault, userDescStyle)
+		desc = selectionHighlightStyle + desc
+	}
+
+	compDescStyle := color.UnquoteRC(e.config.GetString("completion-description-style"))
+
+	return compDescStyle + desc + color.Reset + padded
 }
 
 // cropCompletions - When the user cycles through a completion list longer
