@@ -1,32 +1,34 @@
 package completion
 
 import (
-	"fmt"
-	"regexp"
+	"math"
 	"sort"
+	"strconv"
 	"strings"
-	"unicode/utf8"
 
 	"github.com/reeflective/readline/internal/color"
 	"github.com/reeflective/readline/internal/term"
+	"golang.org/x/exp/slices"
 )
 
 // group is used to structure different types of completions with different
 // display types, autosuffix removal matchers, under their tag heading.
 type group struct {
-	tag           string        // Printed on top of the group's completions
-	values        [][]Candidate // Values are grouped by aliases/rows, with computed paddings.
-	noSpace       SuffixMatcher // Suffixes to remove if a space or non-nil character is entered after the completion.
-	columnsWidth  []int         // Computed width for each column of completions, when aliases
-	listSeparator string        // This is used to separate completion candidates from their descriptions.
-	list          bool          // Force completions to be listed instead of grided
-	noSort        bool          // Don't sort completions
-	aliased       bool          // Are their aliased completions
-	isCurrent     bool          // Currently cycling through this group, for highlighting choice
-	maxLength     int           // Each group can be limited in the number of comps offered
-	tcMaxLength   int           // Used when display is map/list, for determining message width
-	maxDescWidth  int
-	maxCellLength int
+	tag               string        // Printed on top of the group's completions
+	rows              [][]Candidate // Values are grouped by aliases/rows, with computed paddings.
+	noSpace           SuffixMatcher // Suffixes to remove if a space or non-nil character is entered after the completion.
+	columnsWidth      []int         // Computed width for each column of completions, when aliases
+	descriptionsWidth []int         // Computed width for each column of completions, when aliases
+	listSeparator     string        // This is used to separate completion candidates from their descriptions.
+	list              bool          // Force completions to be listed instead of grided
+	noSort            bool          // Don't sort completions
+	aliased           bool          // Are their aliased completions
+	preserveEscapes   bool          // Preserve escape sequences in the completion inserted values.
+	isCurrent         bool          // Currently cycling through this group, for highlighting choice
+	longestValue      int           // Used when display is map/list, for determining message width
+	longestDesc       int           // Used to know how much descriptions can use when there are aliases.
+	maxDescAllowed    int           // Maximum ALLOWED description width.
+	termWidth         int           // Term size queried at beginning of computes by the engine.
 
 	// Selectors (position/bounds) management
 	posX int
@@ -35,173 +37,62 @@ type group struct {
 	maxY int
 }
 
-func (e *Engine) group(comps Values) {
-	// Compute hints once we found either any errors,
-	// or if we have no completions but usage strings.
-	defer func() {
-		e.hintCompletions(comps)
-	}()
-
-	// Nothing else to do if no completions
-	if len(comps.values) == 0 {
-		return
-	}
-
-	// Apply the prefix to the completions, and filter out any
-	// completions that don't match, optionally ignoring case.
-	matchCase := e.config.GetBool("completion-ignore-case")
-	comps.values = comps.values.FilterPrefix(e.prefix, !matchCase)
-
-	comps.values.EachTag(func(tag string, values RawValues) {
-		// Separate the completions that have a description and
-		// those which don't, and devise if there are aliases.
-		vals, noDescVals, aliased := e.groupValues(&comps, values)
-
-		// Create a "first" group with the "first" grouped values
-		e.newGroup(comps, tag, vals, aliased)
-
-		// If we have a remaining group of values without descriptions,
-		// we will print and use them in a separate, anonymous group.
-		if len(noDescVals) > 0 {
-			e.newGroup(comps, "", noDescVals, false)
-		}
-	})
-
-	e.justifyGroups(comps)
-}
-
-// groupValues separates values based on whether they have descriptions, or are aliases of each other.
-func (e *Engine) groupValues(comps *Values, values RawValues) (vals, noDescVals RawValues, aliased bool) {
-	var descriptions []string
-
-	prefix := ""
-	if e.prefix != "\"\"" && e.prefix != "''" {
-		prefix = e.prefix
-	}
-
-	for _, val := range values {
-		// Ensure all values have a display string.
-		if val.Display == "" {
-			val.Display = val.Value
-		}
-
-		// Currently this is because errors are passed as completions.
-		if strings.HasPrefix(val.Value, prefix+"ERR") && val.Value == prefix+"_" {
-			if val.Description != "" && comps != nil {
-				comps.Messages.Add(color.FgRed + val.Description)
-			}
-
-			continue
-		}
-
-		// Grid completions
-		if val.Description == "" {
-			noDescVals = append(noDescVals, val)
-
-			continue
-		}
-
-		// List/map completions.
-		if stringInSlice(val.Description, descriptions) {
-			aliased = true
-		}
-
-		descriptions = append(descriptions, val.Description)
-		vals = append(vals, val)
-	}
-
-	// if no candidates have a description, swap
-	if len(vals) == 0 {
-		vals = noDescVals
-		noDescVals = make(RawValues, 0)
-	}
-
-	return vals, noDescVals, aliased
-}
-
-func (e *Engine) justifyGroups(values Values) {
-	commandGroups := make([]*group, 0)
-	maxCellLength := 0
-
-	for _, group := range e.groups {
-		// Skip groups that are not to be justified
-		if _, justify := values.Pad[group.tag]; !justify {
-			if _, all := values.Pad["*"]; !all {
-				continue
-			}
-		}
-
-		// Skip groups that are aliased or have more than one column
-		if group.aliased || len(group.columnsWidth) > 1 {
-			continue
-		}
-
-		commandGroups = append(commandGroups, group)
-
-		if group.tcMaxLength > maxCellLength {
-			maxCellLength = group.tcMaxLength
-		}
-	}
-
-	for _, group := range commandGroups {
-		group.tcMaxLength = maxCellLength
-	}
-}
-
-func (e *Engine) newGroup(comps Values, tag string, vals RawValues, aliased bool) {
+// newCompletionGroup initializes a group of completions to be displayed in the same area/header.
+func (e *Engine) newCompletionGroup(comps Values, tag string, vals RawValues, descriptions []string) {
 	grp := &group{
-		tag:           tag,
-		noSpace:       comps.NoSpace,
-		listSeparator: "--",
-		posX:          -1,
-		posY:          -1,
-		aliased:       aliased,
-		columnsWidth:  []int{0},
+		tag:          tag,
+		noSpace:      comps.NoSpace,
+		posX:         -1,
+		posY:         -1,
+		columnsWidth: []int{0},
+		termWidth:    term.GetWidth(),
+		longestDesc:  longest(descriptions, true),
 	}
 
-	// Check that all comps have a display value,
-	// and begin computing some parameters.
-	vals = grp.checkDisplays(vals)
+	// Initialize all options for the group.
+	grp.initOptions(e, &comps, tag, vals)
 
-	// Set sorting options, various display styles, etc.
-	grp.setOptions(comps, tag, vals)
+	// Global actions to take on all values.
+	if !grp.noSort {
+		sort.Stable(vals)
+	}
 
-	// Keep computing/devising some parameters and constraints.
-	// This does not do much when we have aliased completions.
-	grp.computeCells(e, vals)
+	// Initial processing of our assigned values:
+	// Compute color/no-color sizes, some max/min, etc.
+	grp.prepareValues(vals)
 
-	// Rearrange all candidates into a matrix of completions,
-	// based on most parameters computed above.
-	grp.makeMatrix(vals)
+	// Generate the full grid of completions.
+	// Special processing is needed when some values
+	// share a common description, they are "aliased".
+	if completionsAreAliases(vals) {
+		grp.initCompletionAliased(vals)
+	} else {
+		grp.initCompletionsGrid(vals)
+	}
 
 	e.groups = append(e.groups, grp)
 }
 
-func (g *group) checkDisplays(vals RawValues) RawValues {
-	if g.aliased {
-		return vals
-	}
-
-	if len(vals) == 0 {
-		g.columnsWidth[0] = term.GetWidth() - 1
-	}
-
-	// Otherwise update the size of the longest candidate
-	for _, val := range vals {
-		valLen := utf8.RuneCountInString(val.Display)
-		if valLen > g.columnsWidth[0] {
-			g.columnsWidth[0] = valLen
-		}
-	}
-
-	return vals
-}
-
-func (g *group) setOptions(comps Values, tag string, vals RawValues) {
+// initOptions checks for global or group-specific options (display, behavior, grouping, etc).
+func (g *group) initOptions(eng *Engine, comps *Values, tag string, vals RawValues) {
 	// Override grid/list displays
 	_, g.list = comps.ListLong[tag]
 	if _, all := comps.ListLong["*"]; all && len(comps.ListLong) == 1 {
 		g.list = true
+	}
+
+	// Description list separator
+	listSep, err := strconv.Unquote(eng.config.GetString("completion-list-separator"))
+	if err != nil {
+		g.listSeparator = "--"
+	} else {
+		g.listSeparator = listSep
+	}
+
+	// Strip escaped characters in the value component.
+	g.preserveEscapes = comps.Escapes[g.tag]
+	if !g.preserveEscapes {
+		g.preserveEscapes = comps.Escapes["*"]
 	}
 
 	// Always list long commands when they have descriptions.
@@ -220,149 +111,319 @@ func (g *group) setOptions(comps Values, tag string, vals RawValues) {
 	}
 
 	// Override sorting or sort if needed
-	_, g.noSort = comps.NoSort[tag]
-	if _, all := comps.NoSort["*"]; all && len(comps.NoSort) == 1 {
+	g.noSort = comps.NoSort[tag]
+	if noSort, all := comps.NoSort["*"]; noSort && all && len(comps.NoSort) == 1 {
 		g.noSort = true
 	}
-
-	if !g.noSort {
-		sort.Slice(vals, func(i, j int) bool {
-			return vals[i].Display < vals[j].Display
-		})
-	}
 }
 
-func (g *group) computeCells(eng *Engine, vals RawValues) {
-	// Aliases will compute themselves individually, later.
-	if g.aliased {
+// initCompletionsGrid arranges completions when there are no aliases.
+func (g *group) initCompletionsGrid(comps RawValues) {
+	if len(comps) == 0 {
 		return
 	}
 
-	if len(vals) == 0 {
-		return
+	pairLength := g.longestValueDescribed(comps)
+	if pairLength > g.termWidth {
+		pairLength = g.termWidth
 	}
 
-	g.tcMaxLength = g.columnsWidth[0]
-
-	// Each value first computes the total amount of space
-	// it is going to take in a row (including the description)
-	for _, val := range vals {
-		candidate := g.displayTrimmed(color.Strip(val.Display))
-		pad := strings.Repeat(" ", g.tcMaxLength-len(candidate))
-		desc := g.descriptionTrimmed(val.Description)
-		display := fmt.Sprintf("%s%s%s", candidate, pad+" ", desc)
-		valLen := utf8.RuneCountInString(display)
-
-		if valLen > g.maxCellLength {
-			g.maxCellLength = valLen
-		}
+	maxColumns := g.termWidth / pairLength
+	if g.list || maxColumns < 0 {
+		maxColumns = 1
 	}
 
-	// Adapt the maximum cell size based on inputrc settings.
-	compDisplayWidth := g.setMaxCellLength(eng)
+	rowCount := int(math.Ceil(float64(len(comps)) / (float64(maxColumns))))
 
-	// We now have the length of the longest completion for this group,
-	// so we devise how many columns we can use to display completions.
-	g.setColumnsWidth(&vals, compDisplayWidth)
+	g.rows = createGrid(comps, rowCount, maxColumns)
+	g.calculateMaxColumnWidths(g.rows)
 }
 
-func (g *group) setMaxCellLength(eng *Engine) int {
-	termWidth := term.GetWidth()
+// initCompletionsGrid arranges completions when some of them share the same description.
+func (g *group) initCompletionAliased(domains []Candidate) {
+	g.aliased = true
 
-	compDisplayWidth := eng.config.GetInt("completion-display-width")
-	if compDisplayWidth > termWidth {
-		compDisplayWidth = -1
-	}
+	// Filter out all duplicates: group aliased completions together.
+	grid, descriptions := g.createDescribedRows(domains)
+	g.calculateMaxColumnWidths(grid)
+	g.wrapExcessAliases(grid, descriptions)
 
-	if compDisplayWidth > 0 && compDisplayWidth < termWidth {
-		if g.maxCellLength < compDisplayWidth {
-			g.maxCellLength = compDisplayWidth
+	g.maxY = len(g.rows)
+	g.maxX = len(g.columnsWidth)
+}
+
+// This createDescribedRows function takes a list of values, a list of descriptions, and the
+// terminal width as input, and returns a list of rows based on the provided requirements:.
+func (g *group) createDescribedRows(values []Candidate) ([][]Candidate, []string) {
+	descriptionMap := make(map[string][]Candidate)
+	uniqueDescriptions := make([]string, 0)
+	rows := make([][]Candidate, 0)
+
+	// Separate duplicates and store them.
+	for i, description := range values {
+		if slices.Contains(uniqueDescriptions, description.Description) {
+			descriptionMap[description.Description] = append(descriptionMap[description.Description], values[i])
 		} else {
-			g.maxCellLength = termWidth
+			uniqueDescriptions = append(uniqueDescriptions, description.Description)
+			descriptionMap[description.Description] = []Candidate{values[i]}
 		}
 	}
 
-	return compDisplayWidth
+	// Sorting helps with easier grids.
+	for _, description := range uniqueDescriptions {
+		row := descriptionMap[description]
+		// slices.Sort(row)
+		// slices.Reverse(row)
+		rows = append(rows, row)
+	}
+
+	return rows, uniqueDescriptions
 }
 
-func (g *group) setColumnsWidth(vals *RawValues, compDisplayWidth int) {
-	g.maxX = term.GetWidth() / (g.maxCellLength)
-	if g.maxX < 1 {
-		g.maxX = 1 // avoid a divide by zero error
+// Wraps all rows for which there are too many aliases to be displayed on a single one.
+func (g *group) wrapExcessAliases(grid [][]Candidate, descriptions []string) {
+	breakeven := 0
+	maxColumns := len(g.columnsWidth)
+
+	for i, width := range g.columnsWidth {
+		if (breakeven + width + 1) > g.termWidth/2 {
+			maxColumns = i
+			break
+		}
+
+		breakeven += width + 1
 	}
 
-	if g.maxX > len(*vals) {
-		g.maxX = len(*vals)
+	var rows [][]Candidate
+
+	for rowIndex := range grid {
+		row := grid[rowIndex]
+
+		for len(row) > maxColumns {
+			rows = append(rows, row[:maxColumns])
+			row = row[maxColumns:]
+		}
+
+		rows = append(rows, row)
 	}
 
-	if g.list || compDisplayWidth == 0 {
-		g.maxX = 1
-	}
-
-	if g.maxX > compDisplayWidth && compDisplayWidth > 0 {
-		g.maxX = compDisplayWidth
-	}
-
-	// We also have the width for each column
-	g.columnsWidth = make([]int, g.maxX)
-	for i := 0; i < g.maxX; i++ {
-		g.columnsWidth[i] = g.maxCellLength
-	}
+	g.rows = rows
+	g.columnsWidth = g.columnsWidth[:maxColumns]
 }
 
-func (g *group) makeMatrix(vals RawValues) {
-nextValue:
-	for _, val := range vals {
-		valLen := utf8.RuneCountInString(val.Display)
+// prepareValues ensures all of them have a display, and starts
+// gathering information on longest/shortest values, etc.
+func (g *group) prepareValues(vals RawValues) RawValues {
+	for pos, value := range vals {
+		if value.Display == "" {
+			value.Display = value.Value
+		}
 
-		// If we have an alias, and we must get the right
-		// column and the right padding for this column.
-		if g.aliased {
-			for i, row := range g.values {
-				if row[0].Description == val.Description {
-					g.values[i] = append(row, val)
-					g.columnsWidth = getColumnPad(g.columnsWidth, valLen, len(g.values[i]))
+		// Only pass for colors regex should be here.
+		value.displayLen = len(color.Strip(value.Display))
+		value.descLen = len(color.Strip(value.Description))
 
-					continue nextValue
-				}
+		if value.displayLen > g.longestValue {
+			g.longestValue = value.displayLen
+		}
+
+		if value.descLen > g.longestDesc {
+			g.longestDesc = value.descLen
+		}
+
+		vals[pos] = value
+	}
+
+	return vals
+}
+
+func (g *group) setMaximumSizes(col int) int {
+	// Get the length of the longest description in the same column.
+	maxDescLen := g.descriptionsWidth[col]
+	valuesRealLen := sum(g.columnsWidth) + len(g.columnsWidth) + len(g.listSep())
+
+	if valuesRealLen+maxDescLen > g.termWidth {
+		maxDescLen = g.termWidth - valuesRealLen
+	} else if valuesRealLen+maxDescLen < g.termWidth {
+		maxDescLen = g.termWidth - valuesRealLen
+	}
+
+	return maxDescLen
+}
+
+// calculateMaxColumnWidths is in charge of optimizing the sizes of rows/columns.
+func (g *group) calculateMaxColumnWidths(grid [][]Candidate) {
+	var numColumns int
+
+	// Get the row with the greatest number of columns.
+	for _, row := range grid {
+		if len(row) > numColumns {
+			numColumns = len(row)
+		}
+	}
+
+	// First, all columns are as wide as the longest of their elements,
+	// regardless of if this longest element is longer than terminal
+	values := make([]int, numColumns)
+	descriptions := make([]int, numColumns)
+
+	for _, row := range grid {
+		for columnIndex, value := range row {
+			if value.displayLen+1 > values[columnIndex] {
+				values[columnIndex] = value.displayLen + 1
+			}
+
+			if value.descLen+1 > descriptions[columnIndex] {
+				descriptions[columnIndex] = value.descLen + 1
 			}
 		}
+	}
 
-		// Else, either add it to the current row if there is still room
-		// on it for this candidate, or add a new one. We only do that when
-		// we know we don't have aliases, or when we don't have to display list.
-		if !g.aliased && g.canFitInRow(term.GetWidth()) && !g.list {
-			g.values[len(g.values)-1] = append(g.values[len(g.values)-1], val)
-		} else {
-			// Else create a new row, and update the row pad.
-			g.values = append(g.values, []Candidate{val})
-			if g.columnsWidth[0] < valLen+1 {
-				g.columnsWidth[0] = valLen + 1
-			}
+	// If we have only one row, it means that the number of columns
+	// multiplied by the size on the longest one will fit into the
+	// terminal, so we can just
+	if len(grid) == 1 && len(grid[0]) <= numColumns && sum(descriptions) == 0 {
+		for i := range values {
+			values[i] = g.longestValue
 		}
 	}
 
-	if g.aliased {
-		g.maxX = len(g.columnsWidth)
-		g.tcMaxLength = sum(g.columnsWidth) + len(g.columnsWidth) + 1
+	// Last time adjustment: try to reallocate any space modulo to each column.
+	shouldPad := len(grid) > 1 && numColumns > 1 && sum(descriptions) == 0
+	intraColumnSpace := (numColumns * 2)
+	totalSpaceUsed := sum(values) + sum(descriptions) + intraColumnSpace
+	freeSpace := g.termWidth - totalSpaceUsed
+
+	if shouldPad && !g.aliased && freeSpace >= numColumns {
+		each := freeSpace / numColumns
+
+		for i := range values {
+			values[i] += each
+		}
 	}
 
-	g.maxY = len(g.values)
-	if g.maxY > g.maxLength && g.maxLength != 0 {
-		g.maxY = g.maxLength
-	}
+	// The group is mostly ready to print and select its values for completion.
+	g.maxY = len(g.rows)
+	g.maxX = len(values)
+	g.columnsWidth = values
+	g.descriptionsWidth = descriptions
 }
 
-func (g *group) canFitInRow(termWidth int) bool {
-	if len(g.values) == 0 {
-		return false
+func (g *group) longestValueDescribed(vals []Candidate) int {
+	var longestDesc, longestVal int
+
+	// Equivalent to `<completion> -- <Description>`,
+	// asuuming no trailing spaces in the completion
+	// nor leading spaces in the description.
+	descSeparatorLen := 1 + len(g.listSeparator) + 1
+
+	// Get the length of the longest value
+	// and the length of the longest description.
+	for _, val := range vals {
+		if val.displayLen > longestVal {
+			longestVal = val.displayLen
+		}
+
+		if val.descLen > longestDesc {
+			longestDesc = val.descLen
+		}
 	}
 
-	if termWidth/(g.maxCellLength)-1 < len(g.values[len(g.values)-1]) {
-		return false
+	if longestDesc > 0 {
+		longestDesc += descSeparatorLen
 	}
 
-	return true
+	// Always add one: there is at least one space between each column.
+	return longestVal + longestDesc + 1
+}
+
+func (g *group) trimDisplay(comp Candidate, pad, col int) (candidate, padded string) {
+	val := comp.Display
+
+	// No display value means padding.
+	if val == "" {
+		return "", padSpace(pad)
+	}
+
+	// Get the allowed length for this column.
+	// The display can never be longer than terminal width.
+	maxDisplayWidth := g.columnsWidth[col] + 1
+
+	if maxDisplayWidth > g.termWidth {
+		maxDisplayWidth = g.termWidth
+	}
+
+	val = sanitizer.Replace(val)
+
+	if comp.displayLen > maxDisplayWidth {
+		val = color.Trim(val, maxDisplayWidth-trailingValueLen)
+		val += "..." // 3 dots + 1 safety space = -3
+
+		return val, " "
+	}
+
+	return val, padSpace(pad)
+}
+
+func (g *group) trimDesc(val Candidate, pad int) (desc, padded string) {
+	desc = val.Description
+	if desc == "" {
+		return desc, padSpace(pad)
+	}
+
+	// We don't compare against the terminal width:
+	// the correct padding should have been computed
+	// based on the space taken by all candidates
+	// described by our current string.
+	if pad > g.maxDescAllowed {
+		pad = g.maxDescAllowed - val.descLen
+	}
+
+	desc = sanitizer.Replace(desc)
+
+	// Trim the description accounting for escapes.
+	if val.descLen > g.maxDescAllowed && g.maxDescAllowed > 0 {
+		desc = color.Trim(desc, g.maxDescAllowed-trailingDescLen)
+		desc += "..." // 3 dots =  -3
+
+		return g.listSep() + desc, ""
+	}
+
+	if val.descLen+pad > g.maxDescAllowed {
+		pad = g.maxDescAllowed - val.descLen
+	}
+
+	return g.listSep() + desc, padSpace(pad)
+}
+
+func (g *group) getPad(value Candidate, columnIndex int, desc bool) int {
+	columns := g.columnsWidth
+	var valLen = value.displayLen - 1
+
+	if desc {
+		columns = g.descriptionsWidth
+		valLen = value.descLen
+	}
+
+	// Ensure we never longer or even fully equal
+	// to the terminal size: we are not really sure
+	// of where offsets might be set in the code...
+	column := columns[columnIndex]
+	if column > g.termWidth-1 {
+		column = g.termWidth - 1
+	}
+
+	padding := column - valLen
+
+	if padding < 0 {
+		return 0
+	}
+
+	return padding
+}
+
+func (g *group) listSep() string {
+	return g.listSeparator + " "
 }
 
 //
@@ -379,8 +440,8 @@ func (g *group) updateIsearch(eng *Engine) {
 
 	suggs := make([]Candidate, 0)
 
-	for i := range g.values {
-		row := g.values[i]
+	for i := range g.rows {
+		row := g.rows[i]
 
 		for _, val := range row {
 			if eng.IsearchRegex.MatchString(val.Value) {
@@ -392,49 +453,36 @@ func (g *group) updateIsearch(eng *Engine) {
 	}
 
 	// Reset the group parameters
-	g.values = make([][]Candidate, 0)
+	g.rows = make([][]Candidate, 0)
 	g.posX = -1
 	g.posY = -1
-	g.columnsWidth = []int{0}
 
-	// Assign the filtered values: we don't need to check
-	// for a separate set of non-described values, as the
-	// completions have already been triaged when generated.
-	vals, _, aliased := eng.groupValues(nil, suggs)
-	g.aliased = aliased
+	// Initial processing of our assigned values:
+	// Compute color/no-color sizes, some max/min, etc.
+	suggs = g.prepareValues(suggs)
 
-	if len(vals) == 0 {
-		return
-	}
-
-	// And perform the usual initialization routines.
-	vals = g.checkDisplays(vals)
-	g.computeCells(eng, vals)
-	g.makeMatrix(vals)
-}
-
-func (g *group) firstCell() {
-	g.posX = 0
-	g.posY = 0
-}
-
-func (g *group) lastCell() {
-	g.posY = len(g.values) - 1
-	g.posX = len(g.columnsWidth) - 1
-
-	if g.aliased {
-		g.findFirstCandidate(0, -1)
+	// Generate the full grid of completions.
+	// Special processing is needed when some values
+	// share a common description, they are "aliased".
+	if completionsAreAliases(suggs) {
+		g.initCompletionAliased(suggs)
 	} else {
-		g.posX = len(g.values[g.posY]) - 1
+		g.initCompletionsGrid(suggs)
 	}
 }
 
 func (g *group) selected() (comp Candidate) {
+	defer func() {
+		if !g.preserveEscapes {
+			comp.Value = color.Strip(comp.Value)
+		}
+	}()
+
 	if g.posY == -1 || g.posX == -1 {
-		return g.values[0][0]
+		return g.rows[0][0]
 	}
 
-	return g.values[g.posY][g.posX]
+	return g.rows[g.posY][g.posX]
 }
 
 func (g *group) moveSelector(x, y int) (done, next bool) {
@@ -462,7 +510,7 @@ func (g *group) moveSelector(x, y int) (done, next bool) {
 		}
 
 		g.posY--
-		g.posX = len(g.values[g.posY]) - 1
+		g.posX = len(g.rows[g.posY]) - 1
 	}
 
 	// 2) If we are reverse-cycling and currently on the first candidate,
@@ -475,7 +523,7 @@ func (g *group) moveSelector(x, y int) (done, next bool) {
 			return true, false
 		}
 
-		g.posY = len(g.values) - 1
+		g.posY = len(g.rows) - 1
 		g.posX--
 	}
 
@@ -491,7 +539,7 @@ func (g *group) moveSelector(x, y int) (done, next bool) {
 	}
 
 	// 4) If we are on the last column, go to next row or next group
-	if g.posX > len(g.values[g.posY])-1 {
+	if g.posX > len(g.rows[g.posY])-1 {
 		if g.aliased {
 			return g.findFirstCandidate(x, y)
 		}
@@ -513,7 +561,7 @@ func (g *group) moveSelector(x, y int) (done, next bool) {
 // otherwise loop in the direction wished until one is found, or go next/
 // previous column, and so on.
 func (g *group) findFirstCandidate(x, y int) (done, next bool) {
-	for g.posX > len(g.values[g.posY])-1 {
+	for g.posX > len(g.rows[g.posY])-1 {
 		g.posY += y
 		g.posY += x
 
@@ -526,7 +574,7 @@ func (g *group) findFirstCandidate(x, y int) (done, next bool) {
 				return true, false
 			}
 
-			g.posY = len(g.values) - 1
+			g.posY = len(g.rows) - 1
 			g.posX--
 		}
 
@@ -544,198 +592,69 @@ func (g *group) findFirstCandidate(x, y int) (done, next bool) {
 	return
 }
 
-func (g *group) writeComps(eng *Engine) (comp string) {
-	if len(g.values) == 0 {
-		return
-	}
-
-	if g.tag != "" {
-		comp += fmt.Sprintf("%s%s%s %s", color.Bold, color.FgYellow, g.tag, color.Reset) + term.ClearLineAfter + term.NewlineReturn
-		eng.usedY++
-	}
-
-	// Base parameters
-	var columns, rows int
-
-	for range g.values {
-		// Generate the completion string for this row (comp/aliases
-		// and/or descriptions), and apply any styles and isearch
-		// highlighting with pattern replacement,
-		comp += g.writeRow(eng, columns)
-
-		columns++
-		rows++
-
-		if rows > g.maxY {
-			break
-		}
-	}
-
-	eng.usedY += rows
-
-	return comp
+func (g *group) firstCell() {
+	g.posX = 0
+	g.posY = 0
 }
 
-func (g *group) writeRow(eng *Engine, row int) (comp string) {
-	current := g.values[row]
+func (g *group) lastCell() {
+	g.posY = len(g.rows) - 1
+	g.posX = len(g.columnsWidth) - 1
 
-	writeDesc := func(val Candidate, x, y, pad int) string {
-		desc := g.highlightDescription(eng, val, y, x)
-		descPad := g.padDescription(current, val, pad)
-
-		if descPad > 0 {
-			desc += strings.Repeat(" ", descPad)
-		}
-
-		return desc
-	}
-
-	for col, val := range current {
-		// Generate the highlighted candidate with its padding
-		isSelected := row == g.posY && col == g.posX && g.isCurrent
-		cell, pad := g.padCandidate(current, val, col)
-		comp += g.highlightCandidate(eng, val, cell, pad, isSelected) + " "
-
-		// And append the description only if at the end of the row,
-		// or if there are no aliases, in which case all comps are described.
-		if !g.aliased || col == len(current)-1 {
-			comp += writeDesc(val, col, row, len(cell)+len(pad))
-		}
-	}
-
-	comp += term.ClearLineAfter + term.NewlineReturn
-
-	return
-}
-
-func (g *group) highlightCandidate(eng *Engine, val Candidate, cell, pad string, selected bool) (candidate string) {
-	reset := color.Fmt(val.Style)
-	candidate = g.displayTrimmed(val.Display)
-
-	if eng.IsearchRegex != nil && eng.isearchBuf.Len() > 0 {
-		match := eng.IsearchRegex.FindString(candidate)
-		match = color.Fmt(color.Bg+"244") + match + color.Reset + reset
-		candidate = eng.IsearchRegex.ReplaceAllLiteralString(candidate, match)
-	}
-
-	switch {
-	// If the comp is currently selected, overwrite any highlighting already applied.
-	case selected:
-		candidate = color.Fmt(color.Bg+"255") + color.FgBlackBright + g.displayTrimmed(color.Strip(val.Display))
-		if g.aliased {
-			candidate += cell + color.Reset
-		}
-
-	default:
-		// Highlight the prefix if any and configured for it.
-		if eng.config.GetBool("colored-completion-prefix") && eng.prefix != "" {
-			if prefixMatch, err := regexp.Compile(fmt.Sprintf("^%s", eng.prefix)); err == nil {
-				candidate = prefixMatch.ReplaceAllString(candidate, color.Bold+color.FgBlue+eng.prefix+color.BoldReset+color.FgDefault+reset)
-			}
-		}
-
-		candidate = reset + candidate + color.Reset + cell
-	}
-
-	return candidate + pad
-}
-
-func (g *group) highlightDescription(eng *Engine, val Candidate, row, col int) (desc string) {
-	if val.Description == "" {
-		return color.Reset
-	}
-
-	desc = g.descriptionTrimmed(val.Description)
-
-	if eng.IsearchRegex != nil && eng.isearchBuf.Len() > 0 {
-		match := eng.IsearchRegex.FindString(desc)
-		match = color.Fmt(color.Bg+"244") + match + color.Reset + color.Dim
-		desc = eng.IsearchRegex.ReplaceAllLiteralString(desc, match)
-	}
-
-	// If the comp is currently selected, overwrite any highlighting already applied.
-	if row == g.posY && col == g.posX && g.isCurrent && !g.aliased {
-		desc = color.Fmt(color.Bg+"255") + color.FgBlackBright + g.descriptionTrimmed(val.Description)
-	}
-
-	return color.Dim + desc + color.Reset
-}
-
-func (g *group) padCandidate(row []Candidate, val Candidate, col int) (cell, pad string) {
-	var cellLen, padLen int
-	valLen := utf8.RuneCountInString(val.Display)
-
-	if !g.aliased {
-		padLen = g.tcMaxLength - valLen
-		if padLen < 0 {
-			padLen = 0
-		}
-
-		return "", strings.Repeat(" ", padLen)
-	}
-
-	cellLen = g.columnsWidth[col] - valLen
-
-	if len(row) == 1 {
-		padLen = sum(g.columnsWidth) + len(g.columnsWidth) - g.columnsWidth[col] - 1
-	}
-
-	return strings.Repeat(" ", cellLen), strings.Repeat(" ", padLen)
-}
-
-func (g *group) padDescription(row []Candidate, val Candidate, valPad int) (pad int) {
 	if g.aliased {
-		return 1
+		g.findFirstCandidate(0, -1)
+	} else {
+		g.posX = len(g.rows[g.posY]) - 1
 	}
-
-	candidateLen := len(g.displayTrimmed(val.Display)) + valPad + 1
-	individualRest := (term.GetWidth() % g.maxCellLength) / (g.maxX + len(row))
-	pad = g.maxCellLength - candidateLen - len(g.descriptionTrimmed(val.Description)) + individualRest
-
-	if pad > 1 {
-		pad--
-	}
-
-	return pad
 }
 
-func (g *group) displayTrimmed(val string) string {
-	termWidth := term.GetWidth()
-	if g.tcMaxLength > termWidth-1 {
-		g.tcMaxLength = termWidth - 1
-	}
+func completionsAreAliases(values []Candidate) bool {
+	oddValueMap := make(map[string]bool)
 
-	if len(val) > g.tcMaxLength {
-		val = val[:g.tcMaxLength-3] + "..."
-	}
-
-	val = sanitizer.Replace(val)
-
-	return val
-}
-
-func (g *group) descriptionTrimmed(desc string) string {
-	if desc == "" {
-		return desc
-	}
-
-	termWidth := term.GetWidth()
-	if g.tcMaxLength > termWidth {
-		g.tcMaxLength = termWidth
-	}
-
-	g.maxDescWidth = termWidth - g.tcMaxLength - len(g.listSeparator) - 1
-
-	if len(desc) >= g.maxDescWidth {
-		offset := 4
-		if !g.aliased {
-			offset++
+	for _, value := range values {
+		if value.Description == "" {
+			continue
 		}
 
-		desc = desc[:g.maxDescWidth-offset] + "..."
+		if _, found := oddValueMap[value.Description]; found {
+			return true
+		}
+
+		oddValueMap[value.Description] = true
 	}
 
-	desc = g.listSeparator + " " + sanitizer.Replace(desc)
+	return false
+}
 
-	return desc
+func createGrid(values []Candidate, rowCount, maxColumns int) [][]Candidate {
+	if rowCount < 0 {
+		rowCount = 0
+	}
+
+	grid := make([][]Candidate, rowCount)
+
+	for i := 0; i < rowCount; i++ {
+		grid[i] = createRow(values, maxColumns, i)
+	}
+
+	return grid
+}
+
+func createRow(domains []Candidate, maxColumns, rowIndex int) []Candidate {
+	rowStart := rowIndex * maxColumns
+	rowEnd := (rowIndex + 1) * maxColumns
+
+	if rowEnd > len(domains) {
+		rowEnd = len(domains)
+	}
+
+	return domains[rowStart:rowEnd]
+}
+
+func padSpace(times int) string {
+	if times > 0 {
+		return strings.Repeat(" ", times)
+	}
+
+	return ""
 }
