@@ -1,6 +1,8 @@
 package completion
 
-import "strings"
+import (
+	"strings"
+)
 
 // RawValues is a list of completion candidates.
 type RawValues []Candidate
@@ -62,22 +64,37 @@ func (c RawValues) EachTag(tagF func(tag string, values RawValues)) {
 
 // FilterPrefix filters values with given prefix.
 // If matchCase is false, the filtering is made case-insensitive.
+// This function ensures that all spaces are correctly.
 func (c RawValues) FilterPrefix(prefix string, matchCase bool) RawValues {
+	if prefix == "" {
+		return c
+	}
+
 	filtered := make(RawValues, 0)
 
 	if !matchCase {
 		prefix = strings.ToLower(prefix)
 	}
 
-	for _, r := range c {
-		val := r.Value
+	for _, raw := range c {
+		val := raw.Value
+
 		if !matchCase {
 			val = strings.ToLower(val)
 		}
 
 		if strings.HasPrefix(val, prefix) {
-			filtered = append(filtered, r)
+			filtered = append(filtered, raw)
 		}
 	}
+
 	return filtered
+}
+
+func (c RawValues) Len() int { return len(c) }
+
+func (c RawValues) Swap(i, j int) { c[i], c[j] = c[j], c[i] }
+
+func (c RawValues) Less(i, j int) bool {
+	return strings.ToLower(c[i].Value) < strings.ToLower(c[j].Value)
 }
